added check if the file is loaded

+++ forms_php/task1.php
@@ -1,7 +1,7 @@
 <?php
 
 echo "<br>";
-error_reporting(0);
+
 
 // Задача 1.1
 echo "Меня зовут ". $_POST['fname']. " ".$_POST['sname']. "<br>Мне " .$_POST['age'] . "<br> Я писал на языках <br>";
@@ -13,6 +13,14 @@ foreach($_POST['checklist'] as $selected){
 echo "Мой любимый язык ".$_POST['fav_lang']. "<br>";
 echo $_POST['opinion']. "<br>";
 
-echo "Вы загрузили файл " .$_FILES['uploadfile']['name']. "<br>Он весит ".$_FILES['uploadfile']['size']. "байт"; 
-     
+
+if (!file_exists($_FILES['uploadfile']['tmp_name']) || !is_uploaded_file($_FILES['uploadfile']['tmp_name']))
+{
+    echo 'No upload';
+}
+else{
+    echo "Ваш файл был успешно загружен <br>";
+    echo "Вы загрузили файл " .$_FILES['uploadfile']['name']. "<br>Он весит ".$_FILES['uploadfile']['size']. "байт"; 
+}
+
  ?>
